Possible bug?
It must be that since the alternate tree is created with the same parent... when filtering down to leaves it also gets a copy of the data automatically. Then it gets yet another copy explicitly? And if you're tracing down to leaves to vote... you're probably tracing down the path to the alternate's leaves as well.

Is this happening?