Alternates must not predict.
This would seem to be the point of the -999 test. Since it doesn't work, alternates are predicting.

But there is much more to this. As noted earlier, simply disabling alternates doesn't help.

What's happening is that the mainline does fine until drift happens. Then even if tree substitution is turned off, alternates predict. But merely turning off alternate prediction shouldn't be a problem.

Since it is a problem, it shows that the mainline is not behaving like VFDT. It stops learning the moment an alternate is created. Further tests show that it actually stops learning if the alternate is allowed to learn. Examples are filtering down to the alternates leaving the mainline dry.

These tests will follow now.