Spurious weighting turned off
There was no mention of this in the paper. It is simply weighting everything and actually leads to worse performance than VFDT on this test.

Turning it off lets it match VFDT performance. Next, I will enable subtree promotion and get closer to HAT-ADWIN as described in the paper and debugged..