HATADWINOriginal for testing purposes
Now let's add the original source and compare.

Apart from the fact that alternate subtrees of depth are hard to come by, there is yet another reason the bug that alternate nodes vote due to a buggy parent check (-999) work.

Both the root and its alternate have parent "null". When the alternate is split, since it's parent is null, root is set to the alternate split node.

Alternate is set to root, and recovery happens.

Since root substitution happens with an alternate of level 0 in the debugged algorithm, what we have is the alternate is almost immediately better than the root. SO whether it is deliberately substituted or accidentally set to root, it happens within the same epoch keeping error the same.


EvaluatePrequential -l trees.HATADWIN -s (generators.monash.AbruptDriftGenerator -o 0.700002 -p -z 5 -n 5 -v 5 -r 1 -i 0.800001 -b 200000) -i 400000 -f 1000