Some output: bugs or design?
THESE FINDINGS ARE EXTREMELY IMPORTANT

VFDT's forgetting runs one deep.
VFDT keeps attribute observers as well as an observedClassDistribution.

The observedClassDistribution is generated by the observers in the parent.

So the observed class distribution is served up as a prior (since it is used for voting, NaiveBayes voting, etc) to the child by the parent

We shouldn't expect counts to tally

Since the original paper is silent on this design issue... we can just take this to be an unconscious design decision in VFDT-MOA