Bugfix. The observed class distribution is getting updated...
twice. So I'm resetting it to what VFDT has. So that the only change is the node statistics. And in fact when we take out populating node statistics we now match VFDT. The weight thing is to validate, to play around with the fact that lower weights transmitted lead to closer performance to VFDT!! So node statistics are contributing to initial sluggishness for VFDTUnforgetting. Now it's time to test on more test sets.